rename windll to windriver (more accurate since we don't use the dll anymore)
